Fix: LTI needs JWT token..
