+++ controllers/default.py
@@ -193,15 +193,9 @@ def index():
     else:
         # At this point the user has logged in
         # add a jwt cookie for compatibility with bookserver
-        token = _create_access_token(
+        _create_access_token(
             {"sub": auth.user.username}, expires=datetime.timedelta(days=30)
         )
-        # set cookie
-        if token:
-            response.cookies["access_token"] = token
-            response.cookies["access_token"]["expires"] = 24 * 3600 * 30
-            response.cookies["access_token"]["path"] = "/"
-
         # check to see if there is an entry in user_courses for
         # this user,course configuration
         in_db = db(
@@ -637,53 +631,3 @@ def delete():
         auth.logout()  # logout user and redirect to home page
     else:
         redirect(URL("default", "user/profile"))
-
-
-# This function is basically copied from the fastapi_login plugin
-# see `their github repo <https://github.com/MushroomMaula/fastapi_login>`_
-#
-def _create_access_token(data: dict, expires=None, scopes=None) -> bytes:
-    """
-    Helper function to create the encoded access token using
-    the provided secret and the algorithm of the LoginManager instance
-
-    Args:
-        data (dict): The data which should be stored in the token
-        expires (datetime.timedelta):  An optional timedelta in which the token expires.
-            Defaults to 15 minutes
-        scopes (Collection): Optional scopes the token user has access to.
-
-    Returns:
-        The encoded JWT with the data and the expiry. The expiry is
-        available under the 'exp' key
-    """
-
-    to_encode = data.copy()
-
-    if expires:
-        expires_in = datetime.datetime.utcnow() + expires
-    else:
-        # default to 15 minutes expiry times
-        expires_in = datetime.datetime.utcnow() + datetime.timedelta(minutes=15)
-
-    to_encode.update({"exp": expires_in})
-
-    if scopes is not None:
-        unique_scopes = set(scopes)
-        to_encode.update({"scopes": list(unique_scopes)})
-
-    algorithm = "HS256"  # normally set in constructor
-
-    # the secret key value should be set in 1.py as part of the
-    # web2py installation.
-    secret = settings.secret
-
-    try:
-        encoded_jwt = jwt.encode(to_encode, secret, algorithm)
-    except:
-        logger.error(f"failed to create a JWT Token for {to_encode}")
-        if not secret:
-            logger.error("Please set a secret key value in models/1.py")
-        encoded_jwt = None
-    # decode here decodes the byte str to a normal str not the token
-    return encoded_jwt
