+++ models/db.py
@@ -609,3 +609,60 @@ def _validateUser(username, password, fname, lname, email, course_name, line):
         errors.append(f"Email address missing @ on line {line}")
 
     return errors
+
+
+# This function is basically copied from the fastapi_login plugin
+# see `their github repo <https://github.com/MushroomMaula/fastapi_login>`_
+#
+def _create_access_token(data: dict, expires=None, scopes=None) -> bytes:
+    """
+    Helper function to create the encoded access token using
+    the provided secret and the algorithm of the LoginManager instance
+
+    Args:
+        data (dict): The data which should be stored in the token
+        expires (datetime.timedelta):  An optional timedelta in which the token expires.
+            Defaults to 15 minutes
+        scopes (Collection): Optional scopes the token user has access to.
+
+    Returns:
+        The encoded JWT with the data and the expiry. The expiry is
+        available under the 'exp' key
+    """
+
+    to_encode = data.copy()
+
+    if expires:
+        expires_in = datetime.datetime.utcnow() + expires
+    else:
+        # default to 15 minutes expiry times
+        expires_in = datetime.datetime.utcnow() + datetime.timedelta(minutes=15)
+
+    to_encode.update({"exp": expires_in})
+
+    if scopes is not None:
+        unique_scopes = set(scopes)
+        to_encode.update({"scopes": list(unique_scopes)})
+
+    algorithm = "HS256"  # normally set in constructor
+
+    # the secret key value should be set in 1.py as part of the
+    # web2py installation.
+    secret = settings.secret
+
+    try:
+        encoded_jwt = jwt.encode(to_encode, secret, algorithm)
+    except:
+        logger.error(f"failed to create a JWT Token for {to_encode}")
+        if not secret:
+            logger.error("Please set a secret key value in models/1.py")
+        encoded_jwt = None
+
+    # Added Runestone-only code: set cookie
+    if encoded_jwt:
+        response.cookies["access_token"] = encoded_jwt
+        response.cookies["access_token"]["expires"] = 24 * 3600 * 30
+        response.cookies["access_token"]["path"] = "/"
+
+    # decode here decodes the byte str to a normal str not the token
+    return encoded_jwt
